requirements.txt
beautifulsoup4
requests
weaviate-client
lxml
